Introduce form object patter for new entry form
This is refactoring, no new functionality has been introduced. Before we
were using a factory method in the Entry model Entry::create_entry to
create new entries and handle the logic of creating the goals which
belong to it. We've replaced that with a FormObject, EntryForm, which is
responsible for creating an entry and it's dependent goals. This is a
better pattern, since it follows more closely the single responseibility
principle.

Also added is the error handling for when child goals have model level
invalidations. This was made easy since we had introduced the
FormObject.