Get first pass of edit entry feature working
This gets the edit entry feature spec to pass, but has not been
refactored yet. The way that the form object was written 'works' but was
not well suited to be expanded to include new functionality.  I want to
save this intermediate step so that i can reference (potential good blog
post) why it is sub-ideal and compare to the better pattern shown after
the refactor.