Get second pass of edit entry feature working
This refactors the first pass, but doesn't add any new functionality. It
is better but still not great, there are some issues with the scope of
what the form object is doing. I want to keep this intermediate stage
around for reference on bad and better ways to do form objects.