Get third and final pass at edit entry feature working
This is another refactoring of the entry form object, which improves on
the previous refactoring and doesn't add new functionality.

This refactoring pivoted the role of initialize to take in the higher level
abstraction of an entry and only and entry, not the date or description
or goal_descriptions.  The setting of the actual form entries was pushed
to update_attributes for both entry creation and entry edit.