Add write notes feature on entry page
Before users could not write notes on an entry show page, this commit
will allow them to.

Users can edit or create entries on same form within
the entry show page. Another strategy, which was not used, would have been
to lean on the entry_form object to initialize entries with blank notes,
and only allow users to edit notes.

Users need to press a "Save Note" button to persist the notes.