Add complete goal feature
This allows users to click a goal on an entry show page and toggle the
goal from uncompleted to completed, and vice versa.

It introduced some javascript to fire off a fetch request to the
toggle-a-goal endpoint and since we are using capybara webkit for
feature testing, this required adding babel and polyfills for the fetch
api and promises.